🚑 一部のAndroid端末でスキャナーが起動できない
